Fix for windows
